Moved the data modifying test to separate file. The saveApplication() test modified data and caused fetchApplication() and fetchAllTestApplications() tests to fail.
