+++ hakemus-api/src/test/java/fi/vm/sade/haku/oppija/hakemus/it/ApplicationDaoReadIT.java
@@ -1,37 +1,17 @@
 package fi.vm.sade.haku.oppija.hakemus.it;
 
-import java.io.File;
-import java.io.IOException;
-
-import fi.vm.sade.haku.oppija.lomake.domain.ApplicationSystem;
+import fi.vm.sade.haku.oppija.hakemus.domain.Application;
 import fi.vm.sade.haku.oppija.lomake.exception.ApplicationSystemNotFound;
 import fi.vm.sade.haku.testfixtures.MongoFixtureImporter;
-import fi.vm.sade.haku.virkailija.lomakkeenhallinta.util.OppijaConstants;
 import org.junit.Test;
-
-import fi.vm.sade.haku.oppija.hakemus.domain.Application;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
 
+import java.io.IOException;
+
 import static org.junit.Assert.*;
 
-public class ApplicationDaoIT extends IntegrationTestSupport {
-
-    @Test
-    public void saveApplication() {
-        final Application application = getTestApplication();
-        final String capitalEmail = "EmailWithCapitalLetters@test.com";
-        application.setOid("1.2.246.562.11.010101");
-        application.getAnswers().get(OppijaConstants.PHASE_PERSONAL).put(OppijaConstants.ELEMENT_ID_EMAIL, capitalEmail);
-        application.getAnswers().get(OppijaConstants.PHASE_PERSONAL).put(OppijaConstants.ELEMENT_ID_HUOLTAJANSAHKOPOSTI, capitalEmail);
-        final Application savedApplication = save(application);
-        final String savedEmail = savedApplication.getPhaseAnswers(OppijaConstants.PHASE_PERSONAL).get(OppijaConstants.ELEMENT_ID_EMAIL);
-        final String savedHuoltajanEmail = savedApplication.getPhaseAnswers(OppijaConstants.PHASE_PERSONAL).get(OppijaConstants.ELEMENT_ID_EMAIL);
-        assertNotEquals(capitalEmail, savedEmail);
-        assertNotEquals(capitalEmail, savedHuoltajanEmail);
-        assertEquals(capitalEmail.toLowerCase(), savedEmail);
-        assertEquals(capitalEmail.toLowerCase(), savedHuoltajanEmail);
-    }
+public class ApplicationDaoReadIT extends IntegrationTestSupport {
 
     @Test
     public void fetchApplication() throws IOException {
